moved the theme directory and fixed the hover
