+++ src/clients/react-redux-vite/src/App.tsx
@@ -1,6 +1,6 @@
 import { BrowserRouter } from 'react-router-dom';
 import { CssBaseline } from '@mui/material';
-import { AppThemeProvider } from './features/theme/AppThemeProvider';
+import { AppThemeProvider } from './theme/AppThemeProvider';
 import NotificationSnackbar from './features/notifications/notificationSnackbar';
 import AppRouter from './routes/AppRouter';
 import { useAuthInit } from './features/auth/authInit';
