+++ src/clients/react-redux-vite/src/app/store.ts
@@ -4,7 +4,7 @@ import { setupListeners } from "@reduxjs/toolkit/query"
 import { counterSlice } from "../features/counter/counterSlice"
 import { quotesApiSlice } from "../features/quotes/quotesApiSlice"
 import { restaurantsApiSlice } from "../features/restaurants/api/apiSlice"
-import { themeSlice } from "../features/theme/themeSlice"
+import { themeSlice } from "../theme/themeSlice"
 import { errorMiddleware } from "../middleware/errorMiddleware"
 import { notificationSlice } from "../features/notifications/notificationSlice"
 import { authSlice } from "./authSlice"
